Fix get dependency tasks

+++ Gruntfile.js
@@ -618,11 +618,16 @@ module.exports = function (grunt) {
 
 	grunt.registerTask('get_toolbox_dependencies', 'Add VN Toolbox Dependencies to bower.json', function () {
 		var fs = require('fs');
-		var _ = require('lodash');
-		var vnBower = JSON.parse(fs.readFileSync('bower_components/vn-toolbox-common/bower.json', 'utf8'));
-		var origBower = JSON.parse(fs.readFileSync('bower.json', 'utf8'));
-		_.extend(origBower.dependencies,  vnBower.dependencies);
-		fs.writeFileSync('bower.json', JSON.stringify(origBower, undefined, 2), 'utf8');
+		fs.exists('bower_components/vn-toolbox-common/bower.json', function (exists) {
+			if(exists){
+				var _ = require('lodash');
+				var vnBower = JSON.parse(fs.readFileSync('bower_components/vn-toolbox-common/bower.json', 'utf8'));
+				var origBower = JSON.parse(fs.readFileSync('bower.json', 'utf8'));
+				_.extend(origBower.dependencies,  vnBower.dependencies);
+				fs.writeFileSync('bower.json', JSON.stringify(origBower, undefined, 2), 'utf8');
+			}
+		});
+
 	});
 
 	grunt.registerTask('default', function () {
